backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
crewai
crewai[tools]
pydantic
openai
PyPDF2